refactor(runtime): consolidate NSM access and split runtime.go into focused files
Replace the monolithic Migrator with methods on *Runtime and introduce an
Nsm session type that owns a single NSM session for batched PCR ops, so
PCR0 is read once in Init and threaded into KMS instead of each subsystem
opening its own session. Deduplicate the SSM read helpers into stateless
readSSMParam/readSSMParamOptional and drop the redundant *Runtime wrappers.

Carve the HTTP plumbing out of runtime.go into cohesive files:
  - middleware.go: signing/logging middleware, gatedMux, init-gating helpers,
    isGRPCRequest (restored to package scope to fix the broken middleware_test)
  - transport.go: protocol-switch reverse-proxy transport + CORS wrapper
  - nsm.go: Nsm session type and PCR operations
  - acme.go: merge of the former acme_cache.go + ACME directory client

Bind the ACME-served leaf via SetTLSKeyHashIfChanged so the attestation
re-binds (and logs) only on cert rotation rather than on every handshake.

Slim cli/verify.go by delegating attestation verification, TLS pinning, and
key binding to the client package, leaving only the PCR0 upgrade-chain check.

Move verifyPCR31Commitment to attestation.go and reorganize tests
(attestation_test.go, acme_test.go, runtime_test.go) to follow the new layout;
coverage is preserved across the renamed files.
