+++ cli/verify.go
@@ -1,31 +1,21 @@
 package cli
 
 import (
-	"bytes"
 	"context"
-	"crypto/rand"
-	"crypto/sha256"
 	"crypto/tls"
 	"encoding/base64"
 	"encoding/hex"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"os"
 	"strings"
 	"time"
 
 	"github.com/ArkLabsHQ/introspector-enclave/client"
-	"github.com/btcsuite/btcd/btcec/v2/schnorr"
 	"github.com/hf/nitrite"
 	"github.com/spf13/cobra"
 )
-
-// errSigningKeyNotReady is returned when the attestation document's signingKeyHash is
-// all zeros, meaning the app hasn't registered its key with nitriding yet.
-var errSigningKeyNotReady = errors.New("attestation key not yet registered (signingKeyHash is all zeros)")
 
 func verifyCmd() *cobra.Command {
 	cmd := &cobra.Command{
@@ -47,317 +37,89 @@ func verifyCmd() *cobra.Command {
 func runVerify(cmd *cobra.Command, args []string) error {
 	baseURL, _ := cmd.Flags().GetString("base-url")
 	pcr0, _ := cmd.Flags().GetString("expected-pcr0")
+	strictTLS, _ := cmd.Flags().GetBool("strict-tls")
+	doVerifyKey, _ := cmd.Flags().GetBool("verify-attestation-key")
+	waitSec, _ := cmd.Flags().GetInt("wait")
 
 	fmt.Printf("[verify] Verifying %s\n", baseURL)
 
-	strictTLS, _ := cmd.Flags().GetBool("strict-tls")
-	// Bootstrap over an unpinned client; the attestation fetch is the only pre-pin request.
-	bootstrapClient := verifyHTTPClient(!strictTLS)
+	// Delegate attestation verification, TLS pinning, and key binding to the
+	// client package (the same path `enclave curl` uses). The only verify-specific
+	// check is the PCR0 upgrade chain.
+	c, err := client.New(baseURL, client.Options{
+		ExpectedPCR0:   pcr0,
+		StrictTLS:      strictTLS,
+		SkipKeyBinding: !doVerifyKey,
+	})
+	if err != nil {
+		return err
+	}
 
-	waitSec, _ := cmd.Flags().GetInt("wait")
-	doVerifyKey, _ := cmd.Flags().GetBool("verify-attestation-key")
-
-	// runAllChecks verifies the attestation, pins the live cert to the attested
-	// fingerprint, and runs every remaining check over the pinned client.
+	ctx := cmd.Context()
 	runAllChecks := func() error {
-		result, err := verifyAttestation(bootstrapClient, baseURL, pcr0)
+		res, err := c.VerifyAttestation(ctx)
 		if err != nil {
 			return err
 		}
-
-		tlsHash, err := tlsKeyHashFromUserData(result)
-		if err != nil {
-			return err
+		fmt.Println("[verify] Attestation verified: PCR0 + TLS cert pinned to attested fingerprint.")
+		if doVerifyKey && res.AttestationKey != "" {
+			fmt.Printf("[verify] Attestation key binding + response signature verified: %s\n", res.AttestationKey)
 		}
-		pinned, err := client.PinnedHTTPClient(tlsHash, strictTLS)
-		if err != nil {
-			return err
-		}
-		fmt.Println("[verify] TLS cert pinned to attested fingerprint.")
-
-		if doVerifyKey {
-			if err := verifyAttestationKeyBinding(pinned, baseURL, result); err != nil {
-				return err
-			}
-		}
-		if err := verifyPCR0Chain(pinned, baseURL); err != nil {
-			return fmt.Errorf("PCR0 chain: %w", err)
-		}
-		return nil
+		return verifyPCR0Chain(ctx, c)
 	}
 
 	if waitSec > 0 {
-		deadline := time.Now().Add(time.Duration(waitSec) * time.Second)
-		interval := 5 * time.Second
-		attempt := 0
-		for {
-			attempt++
+		const interval = 5 * time.Second
+		start := time.Now()
+		deadline := start.Add(time.Duration(waitSec) * time.Second)
+		for attempt := 1; ; attempt++ {
 			err := runAllChecks()
 			if err == nil {
 				break
 			}
-			if time.Now().After(deadline) {
-				return fmt.Errorf("verification failed after %ds: %w", waitSec, err)
+			remaining := time.Until(deadline)
+			if remaining <= 0 {
+				return fmt.Errorf("verification failed after %s: %w", time.Since(start).Truncate(time.Second), err)
 			}
-			remaining := time.Until(deadline).Truncate(time.Second)
-			fmt.Printf("[verify] Attempt %d failed (%s), retrying... (%s remaining)\n", attempt, shortErr(err), remaining)
-			time.Sleep(interval)
+			// Never sleep past the deadline, so the actual wait matches waitSec.
+			sleep := interval
+			if sleep > remaining {
+				sleep = remaining
+			}
+			fmt.Printf("[verify] Attempt %d failed (%s), retrying... (%s remaining)\n",
+				attempt, shortErr(err), remaining.Truncate(time.Second))
+			time.Sleep(sleep)
 		}
-	} else {
-		if err := runAllChecks(); err != nil {
-			return fmt.Errorf("verification failed: %w", err)
-		}
+	} else if err := runAllChecks(); err != nil {
+		return fmt.Errorf("verification failed: %w", err)
 	}
 
 	fmt.Println("[verify] All checks passed.")
 	return nil
 }
 
-// --- Attestation verification ---
-
-func verifyAttestation(client *http.Client, baseURL, expectedPCR0 string) (*nitrite.Result, error) {
-	nonce := make([]byte, 20)
-	if _, err := rand.Read(nonce); err != nil {
-		return nil, fmt.Errorf("generate nonce: %w", err)
-	}
-	nonceHex := hex.EncodeToString(nonce)
-
-	url := strings.TrimRight(baseURL, "/") + "/enclave/attestation?nonce=" + nonceHex
-	req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		body, _ := io.ReadAll(resp.Body)
-		return nil, fmt.Errorf("attestation status %d: %s", resp.StatusCode, strings.TrimSpace(string(body)))
-	}
-
-	payload, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	docB64 := strings.TrimSpace(string(payload))
-	if strings.HasPrefix(docB64, "{") {
-		var parsed struct {
-			Document string `json:"document"`
-		}
-		if err := json.Unmarshal(payload, &parsed); err == nil && parsed.Document != "" {
-			docB64 = parsed.Document
-		}
-	}
-
-	docBytes, err := base64.StdEncoding.DecodeString(docB64)
-	if err != nil {
-		return nil, fmt.Errorf("decode attestation document: %w", err)
-	}
-
-	result, err := nitrite.Verify(docBytes, nitrite.VerifyOptions{
-		CurrentTime: time.Now(),
-	})
-	if err != nil {
-		if result != nil && result.SignatureOK {
-			fmt.Fprintf(os.Stderr, "[verify] warning: attestation signature OK but validation error: %v\n", err)
-		} else {
-			return nil, err
-		}
-	}
-
-	if result == nil || result.Document == nil {
-		return nil, fmt.Errorf("attestation missing document")
-	}
-
-	expectedNonce, err := hex.DecodeString(nonceHex)
-	if err != nil {
-		return nil, fmt.Errorf("decode nonce: %w", err)
-	}
-	if len(result.Document.Nonce) == 0 {
-		return nil, fmt.Errorf("attestation missing nonce")
-	}
-	if !bytes.Equal(result.Document.Nonce, expectedNonce) {
-		return nil, fmt.Errorf("attestation nonce mismatch")
-	}
-
-	if expectedPCR0 != "" {
-		pcr0, ok := result.Document.PCRs[0]
-		if !ok {
-			return nil, fmt.Errorf("attestation missing PCR0")
-		}
-
-		if !strings.EqualFold(hex.EncodeToString(pcr0), expectedPCR0) {
-			return nil, fmt.Errorf("PCR0 mismatch: expected %s, got %s", expectedPCR0, hex.EncodeToString(pcr0))
-		}
-	}
-
-	fmt.Println("[verify] Attestation document verified.")
-	return result, nil
-}
-
-// --- Attestation key binding ---
-
-// verifyAttestationKeyBinding verifies the enclave's ephemeral attestation key
-// by checking that the pubkey from /v1/enclave-info matches the signingKeyHash in
-// the attestation document's UserData, then verifying a live response signature.
-//
-// UserData format (nitriding v1.4.2):
-//
-//	"sha256:" ++ tlsKeyHash(32) ++ ";" ++ "sha256:" ++ signingKeyHash(32)
-//
-// Total 79 bytes. signingKeyHash is at bytes 47:79.
-func verifyAttestationKeyBinding(client *http.Client, baseURL string, attestResult *nitrite.Result) error {
-	if attestResult == nil || attestResult.Document == nil {
-		return fmt.Errorf("no attestation result to verify against")
-	}
-
-	const (
-		hashPrefix = "sha256:"
-		hashSep    = ";"
-		tlsStart   = len(hashPrefix)
-		tlsEnd     = tlsStart + 32
-		sepStart   = tlsEnd
-		sigKeyPrefix  = sepStart + len(hashSep)
-		sigKeyStart   = sigKeyPrefix + len(hashPrefix)
-		sigKeyEnd     = sigKeyStart + 32
-	)
-
-	userData := attestResult.Document.UserData
-	if len(userData) < sigKeyEnd {
-		fmt.Printf("[verify] UserData too short for signingKeyHash: %d bytes (need %d, enclave may not support attestation key)\n", len(userData), sigKeyEnd)
-		return nil
-	}
-	if !bytes.Equal(userData[:tlsStart], []byte(hashPrefix)) {
-		return fmt.Errorf("UserData missing %q prefix at offset 0 (got %q)", hashPrefix, string(userData[:tlsStart]))
-	}
-	if string(userData[sepStart:sigKeyPrefix]) != hashSep {
-		return fmt.Errorf("UserData missing %q separator at offset %d (got %q)", hashSep, sepStart, string(userData[sepStart:sigKeyPrefix]))
-	}
-	if !bytes.Equal(userData[sigKeyPrefix:sigKeyStart], []byte(hashPrefix)) {
-		return fmt.Errorf("UserData missing %q prefix at offset %d (got %q)", hashPrefix, sigKeyPrefix, string(userData[sigKeyPrefix:sigKeyStart]))
-	}
-	signingKeyHash := userData[sigKeyStart:sigKeyEnd]
-
-	allZero := true
-	for _, b := range signingKeyHash {
-		if b != 0 {
-			allZero = false
-			break
-		}
-	}
-	if allZero {
-		return errSigningKeyNotReady
-	}
-
-	info, err := fetchEnclaveInfo(client, baseURL)
-	if err != nil {
-		return fmt.Errorf("fetch enclave info: %w", err)
-	}
-	if info.AttestationPubkey == "" {
-		return fmt.Errorf("enclave reports no attestation pubkey but signingKeyHash is set")
-	}
-
-	attestPubkeyBytes, err := hex.DecodeString(info.AttestationPubkey)
-	if err != nil {
-		return fmt.Errorf("decode attestation pubkey hex: %w", err)
-	}
-
-	expectedHash := sha256.Sum256(attestPubkeyBytes)
-	if !bytes.Equal(expectedHash[:], signingKeyHash) {
-		return fmt.Errorf("signingKeyHash mismatch: expected SHA256(%s) = %s, got %s",
-			info.AttestationPubkey,
-			hex.EncodeToString(expectedHash[:]),
-			hex.EncodeToString(signingKeyHash))
-	}
-
-	fmt.Printf("[verify] Attestation key binding verified via signingKeyHash: %s\n", info.AttestationPubkey)
-
-	if err := verifyResponseSignature(client, baseURL, info.AttestationPubkey); err != nil {
-		return fmt.Errorf("response signature verification: %w", err)
-	}
-
-	return nil
-}
-
-// verifyResponseSignature fetches /v1/enclave-info and verifies the
-// X-Attestation-Signature header against the attestation pubkey.
-func verifyResponseSignature(client *http.Client, baseURL, attestPubkeyHex string) error {
-	infoURL := strings.TrimRight(baseURL, "/") + "/v1/enclave-info"
-	req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, infoURL, nil)
-	if err != nil {
-		return err
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return fmt.Errorf("read response body: %w", err)
-	}
-
-	sigHex := resp.Header.Get("X-Attestation-Signature")
-	if sigHex == "" {
-		return fmt.Errorf("response missing X-Attestation-Signature header")
-	}
-
-	pubkeyBytes, err := hex.DecodeString(attestPubkeyHex)
-	if err != nil {
-		return fmt.Errorf("decode pubkey: %w", err)
-	}
-	// The attestation pubkey is compressed (33 bytes). Extract x-only (32 bytes)
-	// by dropping the prefix byte.
-	if len(pubkeyBytes) == 33 {
-		pubkeyBytes = pubkeyBytes[1:]
-	}
-	pubkey, err := schnorr.ParsePubKey(pubkeyBytes)
-	if err != nil {
-		return fmt.Errorf("parse attestation pubkey: %w", err)
-	}
-
-	sigBytes, err := hex.DecodeString(sigHex)
-	if err != nil {
-		return fmt.Errorf("decode signature hex: %w", err)
-	}
-	sig, err := schnorr.ParseSignature(sigBytes)
-	if err != nil {
-		return fmt.Errorf("parse signature: %w", err)
-	}
-
-	msgHash := sha256.Sum256(body)
-	if !sig.Verify(msgHash[:], pubkey) {
-		return fmt.Errorf("signature verification failed")
-	}
-
-	fmt.Println("[verify] Response signature verified (X-Attestation-Signature valid).")
-	return nil
-}
-
 // --- PCR0 chain verification ---
 
-// verifyPCR0Chain verifies the cryptographic PCR0 upgrade chain.
-// If the enclave reports a previous PCR0 with an attestation document,
-// it verifies the document against the AWS Nitro root certificate and
-// confirms the PCR0 inside matches the reported previous_pcr0.
-func verifyPCR0Chain(client *http.Client, baseURL string) error {
-	info, err := fetchEnclaveInfo(client, baseURL)
+// verifyPCR0Chain verifies the cryptographic PCR0 upgrade chain. If the enclave
+// reports a previous PCR0 with an attestation document, it verifies that document
+// against the AWS Nitro root and confirms the PCR0 inside matches previous_pcr0.
+func verifyPCR0Chain(ctx context.Context, c *client.Client) error {
+	resp, err := c.Get(ctx, "/v1/enclave-info")
 	if err != nil {
 		return fmt.Errorf("fetch enclave info: %w", err)
+	}
+	var info enclaveInfoResponse
+	if err := json.Unmarshal(resp.Body, &info); err != nil {
+		return fmt.Errorf("decode enclave info: %w", err)
+	}
+	if info.Error != "" {
+		return fmt.Errorf("enclave init error: %s", info.Error)
 	}
 
 	if info.PreviousPCR0 == "genesis" {
 		fmt.Println("[verify] PCR0 chain: genesis (first deployment)")
 		return nil
 	}
-
 	if info.PreviousPCR0Attestation == "" {
 		fmt.Printf("[verify] PCR0 chain: previous_pcr0=%s (no attestation proof — unverified)\n", info.PreviousPCR0)
 		return nil
@@ -368,9 +130,8 @@ func verifyPCR0Chain(client *http.Client, baseURL string) error {
 		return fmt.Errorf("decode PCR0 attestation: %w", err)
 	}
 
-	// Verify attestation document against AWS Nitro root certificate.
-	// Use zero time to skip expiration check — the attestation was generated
-	// at upgrade time, which may have been weeks/months ago.
+	// Zero time skips the expiration check — the attestation was generated at
+	// upgrade time, which may have been weeks/months ago.
 	result, err := nitrite.Verify(attestBytes, nitrite.VerifyOptions{})
 	if err != nil {
 		if result != nil && result.SignatureOK {
@@ -379,7 +140,6 @@ func verifyPCR0Chain(client *http.Client, baseURL string) error {
 			return fmt.Errorf("PCR0 attestation verification failed: %w", err)
 		}
 	}
-
 	if result == nil || result.Document == nil {
 		return fmt.Errorf("PCR0 attestation missing document payload")
 	}
@@ -389,7 +149,6 @@ func verifyPCR0Chain(client *http.Client, baseURL string) error {
 		return fmt.Errorf("PCR0 not found in attestation document")
 	}
 	attestedPCR0 := hex.EncodeToString(pcr0)
-
 	if !strings.EqualFold(attestedPCR0, info.PreviousPCR0) {
 		return fmt.Errorf("PCR0 chain mismatch: reported previous_pcr0=%s, attested=%s",
 			info.PreviousPCR0, attestedPCR0)
@@ -409,66 +168,6 @@ type enclaveInfoResponse struct {
 	Error                   string `json:"error,omitempty"`
 }
 
-func fetchEnclaveInfo(client *http.Client, baseURL string) (*enclaveInfoResponse, error) {
-	infoURL := strings.TrimRight(baseURL, "/") + "/v1/enclave-info"
-	req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, infoURL, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("read enclave info: %w", err)
-	}
-
-	var info enclaveInfoResponse
-	if err := json.Unmarshal(body, &info); err != nil {
-		if resp.StatusCode != http.StatusOK {
-			return nil, fmt.Errorf("status %d: %s", resp.StatusCode, strings.TrimSpace(string(body)))
-		}
-		return nil, fmt.Errorf("decode enclave info: %w", err)
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		if info.Error != "" {
-			return &info, fmt.Errorf("enclave init error: %s", info.Error)
-		}
-		return nil, fmt.Errorf("status %d: %s", resp.StatusCode, strings.TrimSpace(string(body)))
-	}
-
-	if info.Error != "" {
-		return &info, fmt.Errorf("enclave init error: %s", info.Error)
-	}
-	return &info, nil
-}
-
-// tlsKeyHashFromUserData extracts the hex SHA-256 of the enclave's TLS leaf cert
-// from attestation user_data (bytes 7:39), rejecting an all-zero binding.
-func tlsKeyHashFromUserData(result *nitrite.Result) (string, error) {
-	if result == nil || result.Document == nil {
-		return "", fmt.Errorf("no attestation result")
-	}
-	const tlsStart, tlsEnd = len("sha256:"), len("sha256:") + 32
-	ud := result.Document.UserData
-	if len(ud) < tlsEnd {
-		return "", fmt.Errorf("user_data too short for tlsKeyHash (%d bytes)", len(ud))
-	}
-	if string(ud[:tlsStart]) != "sha256:" {
-		return "", fmt.Errorf("user_data missing %q prefix", "sha256:")
-	}
-	h := hex.EncodeToString(ud[tlsStart:tlsEnd])
-	if strings.Trim(h, "0") == "" {
-		return "", fmt.Errorf("attested tlsKeyHash is all-zero (enclave bound no TLS cert)")
-	}
-	return h, nil
-}
-
 func verifyHTTPClient(insecure bool) *http.Client {
 	transport := http.DefaultTransport.(*http.Transport).Clone()
 	if insecure {
@@ -483,19 +182,16 @@ func verifyHTTPClient(insecure bool) *http.Client {
 // shortErr returns a concise error string for retry logging.
 func shortErr(err error) string {
 	s := err.Error()
-	if strings.Contains(s, "TLS handshake timeout") {
+	switch {
+	case strings.Contains(s, "TLS handshake timeout"):
 		return "TLS handshake timeout"
-	}
-	if strings.Contains(s, "connection refused") {
+	case strings.Contains(s, "connection refused"):
 		return "connection refused"
-	}
-	if strings.Contains(s, "i/o timeout") {
+	case strings.Contains(s, "i/o timeout"):
 		return "i/o timeout"
-	}
-	if errors.Is(err, errSigningKeyNotReady) {
+	case strings.Contains(s, "not yet registered"):
 		return "attestation key not yet registered"
-	}
-	if strings.Contains(s, "enclave init error:") {
+	case strings.Contains(s, "enclave init error:"):
 		return s
 	}
 	if len(s) > 60 {
