+++ runtime/attestation.go
@@ -1,14 +1,17 @@
 package runtime
 
 import (
+	"bytes"
 	"crypto/rsa"
 	"crypto/sha256"
+	"crypto/sha512"
 	"crypto/x509"
 	"encoding/base64"
 	"encoding/hex"
 	"fmt"
 	"io"
 	"log/slog"
+	"strings"
 	"sync"
 	"time"
 
@@ -121,6 +124,19 @@ func (a *AttestationHashes) SetTLSKeyHash(h [sha256.Size]byte) {
 	a.mu.Lock()
 	a.tlsKeyHash = h
 	a.mu.Unlock()
+}
+
+// SetTLSKeyHashIfChanged records h and reports whether it differed from the
+// current value. Lets the per-handshake ACME cert wrapper re-bind (and log)
+// only when autocert rotates the leaf.
+func (a *AttestationHashes) SetTLSKeyHashIfChanged(h [sha256.Size]byte) bool {
+	a.mu.Lock()
+	defer a.mu.Unlock()
+	if a.tlsKeyHash == h {
+		return false
+	}
+	a.tlsKeyHash = h
+	return true
 }
 
 // SetSigningKeyHash records the hash of the app's response-signing key.
@@ -299,4 +315,41 @@ func getAttestationDocumentB64() (string, error) {
 	}
 
 	return base64.StdEncoding.EncodeToString(resp.Attestation.Document), nil
+}
+
+// verifyPCR31Commitment: nil iff the previous enclave's attestation document
+// committed to handing off to THIS enclave's PCR0 (by extending PCR31). The doc
+// is verified against the AWS Nitro root, and its PCR0 must match the stored
+// predecessor PCR0, before PCR31 is trusted.
+func verifyPCR31Commitment(attestDocB64, myPCR0, prevPCR0 string) error {
+	attestDoc, err := base64.StdEncoding.DecodeString(attestDocB64)
+	if err != nil {
+		return fmt.Errorf("decode attestation base64: %w", err)
+	}
+	result, err := verifyAttestationDoc(attestDoc)
+	if err != nil {
+		return fmt.Errorf("verify predecessor attestation: %w", err)
+	}
+	// Bind the attestation to the claimed predecessor: a different, still
+	// validly-signed enclave's document must not be substitutable.
+	attestedPCR0, ok := result.Document.PCRs[0]
+	if !ok {
+		return fmt.Errorf("PCR0 not found in attestation document")
+	}
+	if !strings.EqualFold(hex.EncodeToString(attestedPCR0), prevPCR0) {
+		return fmt.Errorf("attested PCR0 does not match stored MigrationPreviousPCR0")
+	}
+	pcr31, ok := result.Document.PCRs[migrationPCRIndex]
+	if !ok {
+		return fmt.Errorf("PCR%d not found in attestation document", migrationPCRIndex)
+	}
+	myPCR0Bytes, err := hex.DecodeString(myPCR0)
+	if err != nil {
+		return fmt.Errorf("decode own PCR0 hex: %w", err)
+	}
+	expected := sha512.Sum384(append(make([]byte, 48), myPCR0Bytes...))
+	if !bytes.Equal(pcr31, expected[:]) {
+		return fmt.Errorf("PCR%d mismatch: previous enclave committed to a different target PCR0", migrationPCRIndex)
+	}
+	return nil
 }
